Satisfy the linter
Seven findings, all in code this branch added:

- errcheck on the deferred temp-file cleanup in writeFileAtomic. Made the
  discard explicit rather than silent — on the error paths there is nothing
  useful to do with a failed cleanup.
- an unnecessary uint32 conversion in ListAudio.
- misspell runs with locale US, and I had written honoured and cancelled.
- unparam: the flowsim fixture helper took a flows argument that never varied.

gofmt, go vet and golangci-lint are all clean.

+++ internal/rtpmanager/server/audio.go
@@ -74,7 +74,7 @@ func (s *Server) ListAudio(ctx context.Context, req *rtpv1.ListAudioRequest) (*r
 			info.Problem = probeErr.Error()
 			info.Playable = false
 		} else {
-			info.SampleRate = uint32(wav.SampleRate)
+			info.SampleRate = wav.SampleRate
 			info.Channels = uint32(wav.NumChannels)
 			info.BitsPerSample = uint32(wav.BitsPerSample)
 			info.DurationMs = wav.DurationMs()
@@ -113,7 +113,7 @@ func depth(rel string) int {
 // inside it, so "../../etc/passwd" in a flow cannot read outside the prompt
 // library.
 //
-// An absolute path is honoured as written. Deployments exist that mount prompts
+// An absolute path is honored as written. Deployments exist that mount prompts
 // somewhere specific and say so in the flow, and today — with the audio
 // directory unread — an absolute path is the ONLY thing that reliably works, so
 // reinterpreting it here would break exactly the people who did the careful
